thenextecho-template/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sqlalchemy
aiosqlite
alembic
celery[redis]
gevent
websockets
requests
ollama
Pillow
moviepy==1.0.3
pexels-api
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
aiofiles
beautifulsoup4
pathlib
